+++ console/src/reducers/index.js
@@ -7,6 +7,7 @@ import {reducers as log} from "../process/log";
 import {reducers as form} from "../process/form";
 import {reducers as wizard} from "../process/wizard";
 import {reducers as portal} from "../process/portal";
+import {reducers as projectList} from "../project/list";
 import {reducers as secret} from "../user/secret";
 import {reducers as modal} from "../shared/Modal";
 
@@ -19,6 +20,7 @@ export default {
     form,
     wizard,
     portal,
+    projectList,
     secret,
     modal
 };