+++ console/src/sagas.js
@@ -8,6 +8,7 @@ import {sagas as processLog} from "./process/log";
 import {sagas as form} from "./process/form";
 import {sagas as wizard} from "./process/wizard";
 import {sagas as portal} from "./process/portal";
+import {sagas as projectList} from "./project/list";
 import {sagas as secret} from "./user/secret";
 
 export default function*(): Generator<*, *, *> {
@@ -20,6 +21,7 @@ export default function*(): Generator<*, *, *> {
         fork(form),
         fork(wizard),
         fork(portal),
+        fork(projectList),
         fork(secret)
     ];
 }